requirements.txt
pytorch-lightning==2.2.5
torch==2.3.0
waymo-open-dataset-tf-2-12-0==1.6.4
tensorflow==2.12.0
wandb==0.17.7
numba
shapely
geopandas
rasterio
ray
pytest
pyinstrument
casadi
pyquaternion
retry
hydra-core 
torch-optimizer
torchaudio
torchmetrics
torchvision
tensorboard
aiofiles 
aioboto3 
flatdict 
adjustText 
loralib 
easydict 
einops_exts
aggdraw